# api/docs/swagger.yaml
definitions:
...
  models.Department:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      work_ends_at:
        type: string
      work_starts_at:
        type: string
    type: object
  models.DepartmentResp:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      updated_at:
        type: string
      work_ends_at:
        type: string
      work_starts_at:
        type: string
    type: object
...
  models.DoctorResp:
    properties:
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      end_work_year:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      password:
        type: string
      phone_number:
        type: string
      salary:
        type: number
      spec_ids:
        items:
          type: integer
        type: array
      start_work_year:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
  models.DoctorUpdateReq:
    properties:
      address:
        type: string
      biography:
        type: string
      birth_date:
        type: string
      department_id:
        type: integer
      end_work_year:
        type: string
      full_name:
        type: string
      id:
        type: string
      salary:
        type: number
      start_work_year:
        type: string
      work_years:
        type: integer
    type: object
...
  models.ListDoctors:
    properties:
      count:
        type: integer
      doctors:
        items:
          $ref: '#/definitions/models.DoctorResp'
        type: array
    type: object
...
  models.Status:
    properties:
      message:
        type: string
    type: object
...
host: localhost:7070
info:
  contact: {}
  description: api-gateway
  title: Clinic system
  version: "1.0"
paths:
...
  /v1/department/{id}:
    get:
      consumes:
      - application/json
      description: Get department
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DepartmentResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get department by id
      tags:
      - Department
  /v1/department/create:
    post:
      consumes:
      - application/json
      description: Create a new department with the provided details
      parameters:
      - description: Create department
        in: body
        name: DepartmentInfo
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DepartmentResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: create department
      tags:
      - Department
  /v1/department/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete department
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Status'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete doctor
      tags:
      - Department
  /v1/department/update/{id}:
    put:
      consumes:
      - application/json
      description: Update department
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      - description: Update Department
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/models.DoctorUpdateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update department
      tags:
      - Department
  /v1/departments/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: get all departments
      parameters:
      - description: page
        in: path
        name: page
        type: string
      - description: limit
        in: path
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ListDoctors'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get departments' list
      tags:
      - Department
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"